Refactor logic/UI model into WidgetView pattern (SOLID) Add Connection Screen from design
