--- lib/screens/login_screen/login_screen_logic.dart
@@ -1,8 +1,0 @@
-import 'login_screen.dart';
-
-class LoginScreenLogic {
-  LoginScreenState screen;
-  LoginScreenLogic(this.screen);
-
-
-}
