+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: zephy_client
 description: Client to interact with Zephy server instances
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: 'none'
 version: 1.0.0+1
 
 environment:
